Rename component error section for alert message section
