+++ src/components/UI/AlertMessageSection/index.ts
@@ -1,0 +1,1 @@
+export { AlertMessageSection } from './AlertMessageSection';
