--- src/components/UI/ErrorSection/index.ts
@@ -1,1 +1,0 @@
-export { ErrorSection } from './ErrorSection';
