+++ src/components/UI/index.ts
@@ -4,4 +4,4 @@ export * from './Records';
 export * from './Error';
 export * from './SelectInput';
 export * from './SpeedDial';
-export * from './ErrorSection';
+export * from './AlertMessageSection';
